Commit 3

+++ App/controllers/ListingController.php
@@ -3,7 +3,9 @@
 namespace App\Controllers;
 
 use Framework\Database;
+use Framework\Session;
 use Framework\Validation;
+use Framework\Authorization;
 
 class ListingController
 {
@@ -18,7 +20,7 @@ class ListingController
     public function index()
     {
 
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
 
         loadView('listings/index', ['listings' => $listings]);
@@ -63,7 +65,8 @@ class ListingController
 
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
+
         $newListingData = array_map('sanitize', $newListingData);
 
         $requiredFiedls = ['title', 'description', 'salary', 'email', 'city', 'state'];
@@ -102,6 +105,7 @@ class ListingController
 
         $query = "INSERT INTO listings ({$fields}) VALUES ({$values})";
         $this->db->query($query, $newListingData);
+        Session::setFlashMessage('success_message', 'Listing created successfully');
 
         redirect('/listings');
     }
@@ -117,21 +121,31 @@ class ListingController
         $params = [
             'id' => $id
         ];
+
+
         $listing = $this->db->query('SELECT * FROM listings WHERE id= :id', $params)->fetch();
+
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
         }
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to delete this listing!');
+            return redirect('/listings/' . $listing->id);
+        }
 
         $this->db->query('DELETE FROM listings WHERE id= :id', $params);
 
-        $_SESSION['success_message'] = 'Listing deleted succesfully';
+
+        Session::setFlashMessage('success_message', 'Listing deleted succesfully');
+
 
         redirect('/listings');
     }
 
     public function edit($params)
     {
+
         $id = $params['id'] ?? '';
 
         $params = [
@@ -143,6 +157,10 @@ class ListingController
         if (!$listing) {
             ErrorController::notFound('Listing Not Found!');
             return;
+        }
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to update this listing!');
+            return redirect('/listings/' . $listing->id);
         }
 
 
@@ -166,6 +184,10 @@ class ListingController
         if (!$listing) {
             ErrorController::notFound('Listing Not Found!');
             return;
+        }
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to update this listing!');
+            return redirect('/listings/' . $listing->id);
         }
 
         $allowedFields = ['title', 'description', 'salary', 'tags', 'company', 'address', 'city', 'state', 'phone', 'email', 'requirements', 'benefits'];
@@ -201,9 +223,34 @@ class ListingController
             $updateQuery = "UPDATE listings SET $updateFields WHERE id = :id";
             $updateValues['id'] = $id;
             $this->db->query($updateQuery, $updateValues);
-            $_SESSION['success_message'] = 'Listing Updated';
+            Session::setFlashMessage('success_message', 'Listing Updated');
+
 
             redirect('/listings/' . $id);
         }
+    }
+
+    /**
+     * Search listings by keywords/location 
+     * 
+     * @return void
+     */
+    public function search()
+    {
+        $keywords = isset($_GET['keywords']) ? trim($_GET['keywords']) : '';
+        $location = isset($_GET['location']) ? trim($_GET['location']) : '';
+
+        $query = "SELECT * FROM listings WHERE (title LIKE :keywords OR description LIKE :keywords OR tags LIKE :keywords OR company LIKE :keywords) AND (city LIKE :location OR state LIKE :location)";
+        $params = [
+            'keywords' => "%{$keywords}%",
+            'location' => "%{$location}%"
+        ];
+        $listings = $this->db->query($query, $params)->fetchAll();
+
+        loadView('/listings/index', [
+            'listings' => $listings,
+            'keywords' => $keywords,
+            'location' => $location
+        ]);
     }
 }
